Fix race condition on alert presence
`expected.alert_is_present()` is always true, because it is a Selenium predicate
(a `Callable` that expects a web-driver).

This currently raises a NoSuchAlertException.
We have to actually call the expectation with a driver, which will check if an alert is present,
or return false (it internally catches this exception), instead of calling switch_to ourselves.

Fixes: 69db96c ("Reorder element checking as to avoid 10s wait for dialogue pop-up")

Signed-off-by: Edwin Török <edwin.torok@cloud.com>