CA-389464: avoid race condition with page load
There is some race condition between page loads and fetching the images: force new instance.
Even though the code waits for certain page elements to appear and be in the correct state, apparently this only works on the first page.
When loading the next page these elements will already be there and the page will be considered already loaded even though it hasn't fully finished loading yet.

This is a "big hammer" approach, followup commits will use tabs instead of an entire new process.

The effect of the race condition is that you see graphs from completely unrelated measurements shown in place of other graphs
(which is a distinct bug from the "duplicate title causing overwrite" bug)

Signed-off-by: Edwin Török <edwin.torok@cloud.com>